backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.2
openpyxl==3.1.2          # For Excel (.xlsx) support
pyarrow==15.0.2           # For Parquet support
google-cloud-storage==2.16.0
google-cloud-logging==3.10.0
scikit-learn==1.4.2       # For prediction models
pydantic==1.10.13         # (FastAPI uses this for data validation)
python-multipart==0.0.9   # In case of form uploads
scipy
numpy
joblib==1.4.2